archsigのdatasetとsnapshot周辺を分割する
