+++ tools/archsig/src/lib.rs
@@ -1,62 +1,30 @@
 use std::collections::{BTreeMap, BTreeSet};
 use std::error::Error;
-use std::fs;
-use std::path::Path;
-
-use serde_json::Value;
-
+mod dataset;
 mod extractor;
+mod github;
 mod graph;
 mod policy;
+mod relation_complexity;
 mod runtime;
 pub mod schema;
+mod snapshot;
 
+pub use dataset::build_empirical_dataset;
 pub use extractor::{extract_sig0, extract_sig0_with_policy, extract_sig0_with_runtime};
+pub use github::{build_pr_metadata_from_github_files, build_pr_metadata_from_github_values};
+pub use relation_complexity::{
+    extract_relation_complexity_observation, extract_relation_complexity_observation_from_file,
+};
 pub use schema::*;
+pub use snapshot::{build_signature_diff_report, build_signature_snapshot_record};
 
-use graph::Graph;
+use dataset::{DATASET_SIGNATURE_AXES, dataset_metric_status, dataset_signature_shape};
 
 #[cfg(test)]
 pub(crate) use extractor::{ParsedImport, parse_imports};
 #[cfg(test)]
 pub(crate) use graph::compute_signature;
-
-const DATASET_SIGNATURE_AXES: [&str; 16] = [
-    "hasCycle",
-    "sccMaxSize",
-    "maxDepth",
-    "fanoutRisk",
-    "boundaryViolationCount",
-    "abstractionViolationCount",
-    "sccExcessSize",
-    "maxFanout",
-    "reachableConeSize",
-    "weightedSccRisk",
-    "projectionSoundnessViolation",
-    "lspViolationCount",
-    "nilpotencyIndex",
-    "runtimePropagation",
-    "relationComplexity",
-    "empiricalChangeCost",
-];
-
-const DATASET_DELTA_AXES: [&str; 15] = [
-    "hasCycle",
-    "sccMaxSize",
-    "maxDepth",
-    "fanoutRisk",
-    "boundaryViolationCount",
-    "abstractionViolationCount",
-    "sccExcessSize",
-    "maxFanout",
-    "reachableConeSize",
-    "weightedSccRisk",
-    "projectionSoundnessViolation",
-    "lspViolationCount",
-    "nilpotencyIndex",
-    "runtimePropagation",
-    "relationComplexity",
-];
 
 pub fn validate_component_universe_report(
     document: &Sig0Document,
@@ -426,139 +394,6 @@ pub fn build_feature_extension_report(
     }
 }
 
-pub fn build_empirical_dataset(
-    before: &Sig0Document,
-    after: &Sig0Document,
-    mut input: EmpiricalDatasetInput,
-    after_commit_role: &str,
-) -> Result<EmpiricalSignatureDatasetV0, Box<dyn Error>> {
-    if !matches!(after_commit_role, "head" | "merge") {
-        return Err(format!("unsupported after commit role: {after_commit_role}").into());
-    }
-
-    let before_snapshot = signature_snapshot(
-        before,
-        GitCommitRef {
-            sha: input.pull_request.base_commit.clone(),
-            role: "base".to_string(),
-        },
-    );
-    let after_sha = match after_commit_role {
-        "head" => input.pull_request.head_commit.clone(),
-        "merge" => input
-            .pull_request
-            .merge_commit
-            .clone()
-            .ok_or("after commit role merge requires pullRequest.mergeCommit")?,
-        _ => unreachable!("after commit role is checked above"),
-    };
-    let after_snapshot = signature_snapshot(
-        after,
-        GitCommitRef {
-            sha: after_sha,
-            role: after_commit_role.to_string(),
-        },
-    );
-    let delta_signature_signed = delta_signature_signed(&before_snapshot, &after_snapshot);
-    let metric_delta_status = metric_delta_status(&before_snapshot, &after_snapshot);
-
-    input.analysis_metadata.signature_after_commit_role = after_commit_role.to_string();
-
-    Ok(EmpiricalSignatureDatasetV0 {
-        schema_version: EMPIRICAL_DATASET_SCHEMA_VERSION.to_string(),
-        repository: input.repository,
-        pull_request: input.pull_request,
-        signature_before: before_snapshot,
-        signature_after: after_snapshot,
-        delta_signature_signed,
-        metric_delta_status,
-        pr_metrics: input.pr_metrics,
-        issue_incident_links: input.issue_incident_links,
-        analysis_metadata: input.analysis_metadata,
-    })
-}
-
-pub fn build_signature_snapshot_record(
-    document: &Sig0Document,
-    validation_report: Option<&ComponentUniverseValidationReport>,
-    input: SnapshotRecordInput,
-) -> SignatureSnapshotStoreRecordV0 {
-    let validation_summary =
-        snapshot_validation_summary(validation_report, input.validation_report_path.clone());
-    let mut exclusion_reasons = Vec::new();
-    match validation_summary.result.as_str() {
-        "fail" => exclusion_reasons.push("validationSummary.result = fail".to_string()),
-        "not_run" => exclusion_reasons.push("validationSummary.result = not_run".to_string()),
-        _ => {}
-    }
-
-    SignatureSnapshotStoreRecordV0 {
-        schema_version: SIGNATURE_SNAPSHOT_STORE_SCHEMA_VERSION.to_string(),
-        repository: input.repository,
-        revision: input.revision,
-        scan: input.scan,
-        extractor: SnapshotExtractorMetadata {
-            name: EXTRACTOR_NAME.to_string(),
-            version: EXTRACTOR_VERSION.to_string(),
-            rule_set_version: RULE_SET_VERSION.to_string(),
-            input_schema_version: document.schema_version.clone(),
-        },
-        policy: SnapshotPolicyMetadata {
-            policy_id: document.policies.policy_id.clone(),
-            schema_version: document.policies.schema_version.clone(),
-            version: input.policy_version,
-            source_path: input.policy_source_path.clone(),
-            content_hash: input.policy_content_hash,
-        },
-        signature: dataset_signature_shape(document),
-        metric_status: dataset_metric_status(document),
-        validation_summary,
-        artifacts: SnapshotArtifacts {
-            extractor_output_path: input.extractor_output_path,
-            validation_report_path: input.validation_report_path,
-            policy_path: input.policy_source_path,
-        },
-        analysis_metadata: SnapshotAnalysisMetadata {
-            excluded_from_primary_diff: !exclusion_reasons.is_empty(),
-            exclusion_reasons,
-            tags: input.tags,
-            notes: input.notes,
-        },
-    }
-}
-
-pub fn build_signature_diff_report(
-    before: &SignatureSnapshotStoreRecordV0,
-    after: &SignatureSnapshotStoreRecordV0,
-    before_document: Option<&Sig0Document>,
-    after_document: Option<&Sig0Document>,
-    pr_metadata: &[EmpiricalDatasetInput],
-) -> SignatureDiffReportV0 {
-    let before_snapshot = store_record_signature_snapshot(before, "before");
-    let after_snapshot = store_record_signature_snapshot(after, "after");
-    let delta_signature_signed = delta_signature_signed(&before_snapshot, &after_snapshot);
-    let metric_delta_status = metric_delta_status(&before_snapshot, &after_snapshot);
-    let (worsened_axes, improved_axes, unchanged_axes, unmeasured_axes) =
-        classify_axis_deltas(before, after, &metric_delta_status);
-    let evidence_diff = snapshot_evidence_diff(before_document, after_document);
-    let attribution = snapshot_diff_attribution(after, &evidence_diff, &worsened_axes, pr_metadata);
-
-    SignatureDiffReportV0 {
-        schema_version: SIGNATURE_DIFF_REPORT_SCHEMA_VERSION.to_string(),
-        before: snapshot_diff_endpoint(before),
-        after: snapshot_diff_endpoint(after),
-        comparison_status: snapshot_comparison_status(before, after),
-        delta_signature_signed,
-        metric_delta_status,
-        worsened_axes,
-        improved_axes,
-        unchanged_axes,
-        unmeasured_axes,
-        evidence_diff,
-        attribution,
-    }
-}
-
 pub fn build_air_document(
     sig0: &Sig0Document,
     validation: Option<&ComponentUniverseValidationReport>,
@@ -794,610 +629,6 @@ pub fn build_air_document(
             split_status: "unmeasured".to_string(),
         },
     }
-}
-
-pub fn build_pr_metadata_from_github_files(
-    pull_request_path: &Path,
-    files_path: &Path,
-    reviews_path: Option<&Path>,
-    review_threads_path: Option<&Path>,
-) -> Result<EmpiricalDatasetInput, Box<dyn Error>> {
-    let pull_request: Value = serde_json::from_str(&fs::read_to_string(pull_request_path)?)?;
-    let files: Value = serde_json::from_str(&fs::read_to_string(files_path)?)?;
-    let reviews = match reviews_path {
-        Some(path) => Some(serde_json::from_str(&fs::read_to_string(path)?)?),
-        None => None,
-    };
-    let review_threads = match review_threads_path {
-        Some(path) => Some(serde_json::from_str(&fs::read_to_string(path)?)?),
-        None => None,
-    };
-
-    build_pr_metadata_from_github_values(
-        &pull_request,
-        &files,
-        reviews.as_ref(),
-        review_threads.as_ref(),
-    )
-}
-
-pub fn build_pr_metadata_from_github_values(
-    pull_request: &Value,
-    files: &Value,
-    reviews: Option<&Value>,
-    review_threads: Option<&Value>,
-) -> Result<EmpiricalDatasetInput, Box<dyn Error>> {
-    let file_items = json_items(
-        files,
-        &["data", "repository", "pullRequest", "files", "nodes"],
-    );
-    let changed_files =
-        optional_usize(pull_request, &["changed_files"]).unwrap_or(file_items.len());
-    let changed_lines_added = optional_usize(pull_request, &["additions"])
-        .unwrap_or_else(|| sum_usize_field(&file_items, &["additions"]));
-    let changed_lines_deleted = optional_usize(pull_request, &["deletions"])
-        .unwrap_or_else(|| sum_usize_field(&file_items, &["deletions"]));
-
-    let review_items = reviews
-        .map(|reviews| {
-            json_items(
-                reviews,
-                &["data", "repository", "pullRequest", "reviews", "nodes"],
-            )
-        })
-        .unwrap_or_default();
-    let thread_items = review_threads
-        .map(|threads| {
-            json_items(
-                threads,
-                &[
-                    "data",
-                    "repository",
-                    "pullRequest",
-                    "reviewThreads",
-                    "nodes",
-                ],
-            )
-        })
-        .unwrap_or_default();
-
-    let created_at = required_string(pull_request, &["created_at"])?;
-    let merged_at = optional_string(pull_request, &["merged_at"]);
-
-    Ok(EmpiricalDatasetInput {
-        repository: RepositoryRef {
-            owner: required_string(pull_request, &["base", "repo", "owner", "login"])?,
-            name: required_string(pull_request, &["base", "repo", "name"])?,
-            default_branch: required_string(pull_request, &["base", "repo", "default_branch"])?,
-        },
-        pull_request: PullRequestRef {
-            number: required_usize(pull_request, &["number"])?,
-            author: required_string(pull_request, &["user", "login"])?,
-            created_at: created_at.clone(),
-            merged_at: merged_at.clone(),
-            base_commit: required_string(pull_request, &["base", "sha"])?,
-            head_commit: required_string(pull_request, &["head", "sha"])?,
-            merge_commit: optional_string(pull_request, &["merge_commit_sha"]),
-            labels: pull_request_labels(pull_request),
-            is_bot_generated: is_bot_generated_pull_request(pull_request),
-        },
-        pr_metrics: PullRequestMetrics {
-            changed_files,
-            changed_lines_added,
-            changed_lines_deleted,
-            changed_components: changed_components_from_github_files(&file_items),
-            review_comment_count: optional_usize(pull_request, &["review_comments"]).unwrap_or(0),
-            review_thread_count: thread_items.len(),
-            review_round_count: review_items
-                .iter()
-                .filter(|review| review_timestamp(review).is_some())
-                .count(),
-            first_review_latency_hours: first_review_latency_hours(&created_at, &review_items),
-            approval_latency_hours: approval_latency_hours(&created_at, &review_items),
-            merge_latency_hours: merged_at
-                .as_deref()
-                .and_then(|merged_at| hours_between(&created_at, merged_at)),
-        },
-        issue_incident_links: Vec::new(),
-        analysis_metadata: AnalysisMetadata::default(),
-    })
-}
-
-pub fn extract_relation_complexity_observation_from_file(
-    path: &Path,
-) -> Result<RelationComplexityObservation, Box<dyn Error>> {
-    let source = fs::read_to_string(path)?;
-    let input: RelationComplexityCandidateFile = serde_json::from_str(&source)?;
-    extract_relation_complexity_observation(input)
-}
-
-pub fn extract_relation_complexity_observation(
-    mut input: RelationComplexityCandidateFile,
-) -> Result<RelationComplexityObservation, Box<dyn Error>> {
-    if input.schema_version != RELATION_COMPLEXITY_CANDIDATE_SCHEMA_VERSION {
-        return Err(format!(
-            "unsupported relation complexity candidate schemaVersion: {}",
-            input.schema_version
-        )
-        .into());
-    }
-
-    input.measurement_universe.rule_set_version =
-        Some(RELATION_COMPLEXITY_RULE_SET_VERSION.to_string());
-
-    let mut excluded_evidence = input.excluded_evidence;
-    let mut included_by_id: BTreeMap<String, RelationComplexityEvidence> = BTreeMap::new();
-    let unsupported_frameworks =
-        unsupported_relation_complexity_frameworks(&input.measurement_universe.frameworks);
-    let framework_supported = input.measurement_universe.frameworks.is_empty()
-        || input
-            .measurement_universe
-            .frameworks
-            .iter()
-            .any(|framework| relation_complexity_framework_supported(framework));
-    let unsupported_reason = (!unsupported_frameworks.is_empty())
-        .then(|| format!("unsupported-framework:{}", unsupported_frameworks.join(",")));
-
-    let candidates = input
-        .evidence
-        .into_iter()
-        .chain(input.evidence_candidates)
-        .collect::<Vec<_>>();
-    for mut candidate in candidates {
-        validate_relation_complexity_candidate_shape(&candidate)?;
-
-        if !framework_supported {
-            excluded_evidence.push(excluded_relation_complexity_evidence(
-                &candidate,
-                unsupported_reason
-                    .as_deref()
-                    .expect("unsupported reason exists"),
-            ));
-            continue;
-        }
-
-        let ownership = candidate.ownership.as_str();
-        if !matches!(ownership, "application-owned" | "application-configured") {
-            excluded_evidence.push(excluded_relation_complexity_evidence(
-                &candidate,
-                &format!("ownership-not-counted:{ownership}"),
-            ));
-            continue;
-        }
-
-        if matches!(
-            candidate.review_status.as_str(),
-            "excluded" | "false-positive" | "rejected"
-        ) {
-            excluded_evidence.push(excluded_relation_complexity_evidence(
-                &candidate,
-                &format!("review-status-not-counted:{}", candidate.review_status),
-            ));
-            continue;
-        }
-
-        let mut unsupported_tags = Vec::new();
-        candidate.tags = relation_complexity_tags(candidate.tags, &mut unsupported_tags);
-        if !unsupported_tags.is_empty() {
-            excluded_evidence.push(excluded_relation_complexity_evidence(
-                &candidate,
-                &format!("unsupported-tags:{}", unsupported_tags.join(",")),
-            ));
-        }
-        if candidate.tags.is_empty() {
-            excluded_evidence.push(excluded_relation_complexity_evidence(
-                &candidate,
-                "no-counted-relation-complexity-tags",
-            ));
-            continue;
-        }
-
-        included_by_id
-            .entry(candidate.id.clone())
-            .and_modify(|existing| {
-                existing.tags = relation_complexity_tags(
-                    existing
-                        .tags
-                        .iter()
-                        .cloned()
-                        .chain(candidate.tags.iter().cloned())
-                        .collect(),
-                    &mut Vec::new(),
-                );
-            })
-            .or_insert(candidate);
-    }
-
-    let evidence: Vec<RelationComplexityEvidence> = included_by_id.into_values().collect();
-    let counts = relation_complexity_counts(&evidence);
-    let relation_complexity = counts.relation_complexity();
-
-    Ok(RelationComplexityObservation {
-        schema_version: RELATION_COMPLEXITY_OBSERVATION_SCHEMA_VERSION.to_string(),
-        repository: input.repository,
-        revision: input.revision,
-        measurement_universe: input.measurement_universe,
-        workflow: input.workflow,
-        counts,
-        relation_complexity,
-        evidence,
-        excluded_evidence,
-    })
-}
-
-fn validate_relation_complexity_candidate_shape(
-    candidate: &RelationComplexityEvidence,
-) -> Result<(), String> {
-    if candidate.id.is_empty() {
-        return Err("relation complexity evidence id must not be empty".to_string());
-    }
-    if candidate.path.is_empty() {
-        return Err(format!(
-            "relation complexity evidence path must not be empty for {}",
-            candidate.id
-        ));
-    }
-    if candidate.ownership.is_empty() {
-        return Err(format!(
-            "relation complexity evidence ownership must not be empty for {}",
-            candidate.id
-        ));
-    }
-    if candidate.review_status.is_empty() {
-        return Err(format!(
-            "relation complexity evidence reviewStatus must not be empty for {}",
-            candidate.id
-        ));
-    }
-    Ok(())
-}
-
-fn relation_complexity_tags(tags: Vec<String>, unsupported_tags: &mut Vec<String>) -> Vec<String> {
-    let mut supported = BTreeSet::new();
-    let mut unsupported = BTreeSet::new();
-    for tag in tags {
-        if relation_complexity_tag_rank(&tag).is_some() {
-            supported.insert(tag);
-        } else {
-            unsupported.insert(tag);
-        }
-    }
-    unsupported_tags.extend(unsupported);
-
-    let mut tags: Vec<String> = supported.into_iter().collect();
-    tags.sort_by_key(|tag| relation_complexity_tag_rank(tag).expect("supported tag"));
-    tags
-}
-
-fn relation_complexity_tag_rank(tag: &str) -> Option<usize> {
-    match tag {
-        "constraints" => Some(0),
-        "compensations" => Some(1),
-        "projections" => Some(2),
-        "failureTransitions" => Some(3),
-        "idempotencyRequirements" => Some(4),
-        _ => None,
-    }
-}
-
-fn relation_complexity_counts(
-    evidence: &[RelationComplexityEvidence],
-) -> RelationComplexityComponents {
-    let mut counts = RelationComplexityComponents {
-        constraints: 0,
-        compensations: 0,
-        projections: 0,
-        failure_transitions: 0,
-        idempotency_requirements: 0,
-    };
-
-    for item in evidence {
-        let tags: BTreeSet<&str> = item.tags.iter().map(String::as_str).collect();
-        for tag in tags {
-            match tag {
-                "constraints" => counts.constraints += 1,
-                "compensations" => counts.compensations += 1,
-                "projections" => counts.projections += 1,
-                "failureTransitions" => counts.failure_transitions += 1,
-                "idempotencyRequirements" => counts.idempotency_requirements += 1,
-                _ => {}
-            }
-        }
-    }
-
-    counts
-}
-
-fn unsupported_relation_complexity_frameworks(frameworks: &[String]) -> Vec<String> {
-    frameworks
-        .iter()
-        .filter(|framework| !relation_complexity_framework_supported(framework))
-        .cloned()
-        .collect()
-}
-
-fn relation_complexity_framework_supported(framework: &str) -> bool {
-    matches!(
-        framework,
-        "generic-workflow" | "event-sourcing" | "saga" | "crud" | "lean-fixture"
-    )
-}
-
-fn excluded_relation_complexity_evidence(
-    evidence: &RelationComplexityEvidence,
-    reason: &str,
-) -> RelationComplexityExcludedEvidence {
-    RelationComplexityExcludedEvidence {
-        path: evidence.path.clone(),
-        reason: reason.to_string(),
-        symbol: evidence.symbol.clone(),
-        line: evidence.line,
-    }
-}
-
-fn json_field<'a>(value: &'a Value, path: &[&str]) -> Option<&'a Value> {
-    path.iter()
-        .try_fold(value, |current, key| current.get(*key))
-}
-
-fn required_string(value: &Value, path: &[&str]) -> Result<String, Box<dyn Error>> {
-    optional_string(value, path)
-        .ok_or_else(|| format!("missing or non-string GitHub field: {}", path.join(".")).into())
-}
-
-fn optional_string(value: &Value, path: &[&str]) -> Option<String> {
-    match json_field(value, path)? {
-        Value::String(text) => Some(text.clone()),
-        Value::Null => None,
-        _ => None,
-    }
-}
-
-fn required_usize(value: &Value, path: &[&str]) -> Result<usize, Box<dyn Error>> {
-    optional_usize(value, path)
-        .ok_or_else(|| format!("missing or non-integer GitHub field: {}", path.join(".")).into())
-}
-
-fn optional_usize(value: &Value, path: &[&str]) -> Option<usize> {
-    json_field(value, path)?
-        .as_u64()
-        .and_then(|value| usize::try_from(value).ok())
-}
-
-fn json_items<'a>(value: &'a Value, graph_ql_nodes_path: &[&str]) -> Vec<&'a Value> {
-    if let Some(items) = value.as_array() {
-        return items.iter().collect();
-    }
-    if let Some(items) = json_field(value, &["nodes"]).and_then(Value::as_array) {
-        return items.iter().collect();
-    }
-    if let Some(items) = json_field(value, graph_ql_nodes_path).and_then(Value::as_array) {
-        return items.iter().collect();
-    }
-    Vec::new()
-}
-
-fn sum_usize_field(items: &[&Value], path: &[&str]) -> usize {
-    items
-        .iter()
-        .filter_map(|item| optional_usize(item, path))
-        .sum()
-}
-
-fn pull_request_labels(pull_request: &Value) -> Vec<String> {
-    let mut labels: Vec<String> = json_field(pull_request, &["labels"])
-        .and_then(Value::as_array)
-        .into_iter()
-        .flatten()
-        .filter_map(|label| optional_string(label, &["name"]))
-        .collect();
-    labels.sort();
-    labels.dedup();
-    labels
-}
-
-fn is_bot_generated_pull_request(pull_request: &Value) -> bool {
-    let user = match json_field(pull_request, &["user"]) {
-        Some(user) => user,
-        None => return false,
-    };
-    let login = optional_string(user, &["login"]).unwrap_or_default();
-    let user_type = optional_string(user, &["type"]).unwrap_or_default();
-    user_type == "Bot" || login.ends_with("[bot]") || login.ends_with("-bot")
-}
-
-fn changed_components_from_github_files(files: &[&Value]) -> Vec<String> {
-    let components: BTreeSet<String> = files
-        .iter()
-        .filter_map(|file| {
-            optional_string(file, &["filename"])
-                .or_else(|| optional_string(file, &["path"]))
-                .or_else(|| optional_string(file, &["name"]))
-        })
-        .filter_map(|path| lean_module_id_from_path(&path))
-        .collect();
-    components.into_iter().collect()
-}
-
-fn lean_module_id_from_path(path: &str) -> Option<String> {
-    let path = path.trim_start_matches("./");
-    let module_path = path.strip_suffix(".lean")?;
-    let module_id = module_path.replace(['/', '\\'], ".");
-    (!module_id.is_empty()).then_some(module_id)
-}
-
-fn review_timestamp(review: &Value) -> Option<String> {
-    optional_string(review, &["submitted_at"]).or_else(|| optional_string(review, &["submittedAt"]))
-}
-
-fn review_state(review: &Value) -> Option<String> {
-    optional_string(review, &["state"]).map(|state| state.to_ascii_uppercase())
-}
-
-fn first_review_latency_hours(created_at: &str, reviews: &[&Value]) -> Option<f64> {
-    reviews
-        .iter()
-        .filter_map(|review| review_timestamp(review))
-        .filter_map(|submitted_at| seconds_between(created_at, &submitted_at))
-        .min()
-        .map(hours_from_seconds)
-}
-
-fn approval_latency_hours(created_at: &str, reviews: &[&Value]) -> Option<f64> {
-    reviews
-        .iter()
-        .filter(|review| review_state(review).as_deref() == Some("APPROVED"))
-        .filter_map(|review| review_timestamp(review))
-        .filter_map(|submitted_at| seconds_between(created_at, &submitted_at))
-        .min()
-        .map(hours_from_seconds)
-}
-
-fn hours_between(start: &str, end: &str) -> Option<f64> {
-    seconds_between(start, end).map(hours_from_seconds)
-}
-
-fn seconds_between(start: &str, end: &str) -> Option<i64> {
-    let start = parse_github_timestamp(start)?;
-    let end = parse_github_timestamp(end)?;
-    (end >= start).then_some(end - start)
-}
-
-fn hours_from_seconds(seconds: i64) -> f64 {
-    seconds as f64 / 3600.0
-}
-
-fn parse_github_timestamp(timestamp: &str) -> Option<i64> {
-    let timestamp = timestamp.strip_suffix('Z')?;
-    let (date, time) = timestamp.split_once('T')?;
-    let mut date_parts = date.split('-');
-    let year: i64 = date_parts.next()?.parse().ok()?;
-    let month: i64 = date_parts.next()?.parse().ok()?;
-    let day: i64 = date_parts.next()?.parse().ok()?;
-    if date_parts.next().is_some() {
-        return None;
-    }
-
-    let time = time.split_once('.').map_or(time, |(whole, _)| whole);
-    let mut time_parts = time.split(':');
-    let hour: i64 = time_parts.next()?.parse().ok()?;
-    let minute: i64 = time_parts.next()?.parse().ok()?;
-    let second: i64 = time_parts.next()?.parse().ok()?;
-    if time_parts.next().is_some()
-        || !(1..=12).contains(&month)
-        || !(1..=31).contains(&day)
-        || !(0..=23).contains(&hour)
-        || !(0..=59).contains(&minute)
-        || !(0..=60).contains(&second)
-    {
-        return None;
-    }
-
-    Some(days_from_civil(year, month, day) * 86_400 + hour * 3_600 + minute * 60 + second)
-}
-
-fn days_from_civil(year: i64, month: i64, day: i64) -> i64 {
-    let year = year - i64::from(month <= 2);
-    let era = if year >= 0 { year } else { year - 399 } / 400;
-    let year_of_era = year - era * 400;
-    let month_prime = month + if month > 2 { -3 } else { 9 };
-    let day_of_year = (153 * month_prime + 2) / 5 + day - 1;
-    let day_of_era = year_of_era * 365 + year_of_era / 4 - year_of_era / 100 + day_of_year;
-    era * 146_097 + day_of_era - 719_468
-}
-
-fn signature_snapshot(document: &Sig0Document, commit: GitCommitRef) -> SignatureSnapshot {
-    let signature = dataset_signature_shape(document);
-    SignatureSnapshot {
-        commit,
-        extractor: ExtractorRef {
-            name: EXTRACTOR_NAME.to_string(),
-            version: EXTRACTOR_VERSION.to_string(),
-            rule_set_version: RULE_SET_VERSION.to_string(),
-            policy_version: document.policies.policy_id.clone(),
-        },
-        metric_status: dataset_metric_status(document),
-        signature,
-    }
-}
-
-fn dataset_signature_shape(document: &Sig0Document) -> ArchitectureSignatureV1DatasetShape {
-    let graph = Graph::from_components_and_edges(&document.components, &document.edges);
-    let runtime_propagation = document
-        .runtime_dependency_graph
-        .as_ref()
-        .map(|projection| {
-            Graph::from_components_and_edges(&document.components, &projection.edges)
-                .reachable_cone_size()
-        });
-
-    ArchitectureSignatureV1DatasetShape {
-        has_cycle: document.signature.has_cycle,
-        scc_max_size: document.signature.scc_max_size,
-        max_depth: document.signature.max_depth,
-        fanout_risk: document.signature.fanout_risk,
-        boundary_violation_count: document.signature.boundary_violation_count,
-        abstraction_violation_count: document.signature.abstraction_violation_count,
-        scc_excess_size: document.signature.scc_max_size.saturating_sub(1),
-        max_fanout: graph.max_fanout(),
-        reachable_cone_size: graph.reachable_cone_size(),
-        weighted_scc_risk: None,
-        projection_soundness_violation: None,
-        lsp_violation_count: None,
-        nilpotency_index: None,
-        runtime_propagation,
-        relation_complexity: None,
-        empirical_change_cost: None,
-    }
-}
-
-fn dataset_metric_status(document: &Sig0Document) -> BTreeMap<String, MetricStatus> {
-    let mut status = BTreeMap::new();
-    for axis in DATASET_SIGNATURE_AXES {
-        let axis_status = match axis {
-            "hasCycle"
-            | "sccMaxSize"
-            | "maxDepth"
-            | "fanoutRisk"
-            | "boundaryViolationCount"
-            | "abstractionViolationCount" => document
-                .metric_status
-                .get(axis)
-                .cloned()
-                .unwrap_or_else(|| {
-                    unmeasured_status(format!("legacy data without metricStatus entry for {axis}"))
-                }),
-            "sccExcessSize" => measured_status("archsig:derived-scc-excess"),
-            "maxFanout" => measured_status("archsig:import-graph"),
-            "reachableConeSize" => measured_status("archsig:import-graph"),
-            "weightedSccRisk" => unmeasured_status("weight rule set not provided".to_string()),
-            "projectionSoundnessViolation" => {
-                unmeasured_status("projection rule set not provided".to_string())
-            }
-            "lspViolationCount" => {
-                unmeasured_status("observation rule set not provided".to_string())
-            }
-            "nilpotencyIndex" => unmeasured_status("matrix bridge output not provided".to_string()),
-            "runtimePropagation" => {
-                document
-                    .metric_status
-                    .get(axis)
-                    .cloned()
-                    .unwrap_or_else(|| {
-                        unmeasured_status("runtime dependency graph not provided".to_string())
-                    })
-            }
-            "relationComplexity" => {
-                unmeasured_status("relation complexity observation not provided".to_string())
-            }
-            "empiricalChangeCost" => {
-                unmeasured_status("empirical target variable not provided".to_string())
-            }
-            _ => unreachable!("all dataset axes are covered"),
-        };
-        status.insert(axis.to_string(), axis_status);
-    }
-    status
 }
 
 fn air_signature_axes(
@@ -2438,612 +1669,6 @@ fn combined_measurement_boundary(boundaries: &[String]) -> String {
     }
 }
 
-fn delta_signature_signed(
-    before: &SignatureSnapshot,
-    after: &SignatureSnapshot,
-) -> NullableSignatureIntVector {
-    NullableSignatureIntVector {
-        has_cycle: signed_delta("hasCycle", before, after),
-        scc_max_size: signed_delta("sccMaxSize", before, after),
-        max_depth: signed_delta("maxDepth", before, after),
-        fanout_risk: signed_delta("fanoutRisk", before, after),
-        boundary_violation_count: signed_delta("boundaryViolationCount", before, after),
-        abstraction_violation_count: signed_delta("abstractionViolationCount", before, after),
-        scc_excess_size: signed_delta("sccExcessSize", before, after),
-        max_fanout: signed_delta("maxFanout", before, after),
-        reachable_cone_size: signed_delta("reachableConeSize", before, after),
-        weighted_scc_risk: signed_delta("weightedSccRisk", before, after),
-        projection_soundness_violation: signed_delta("projectionSoundnessViolation", before, after),
-        lsp_violation_count: signed_delta("lspViolationCount", before, after),
-        nilpotency_index: signed_delta("nilpotencyIndex", before, after),
-        runtime_propagation: signed_delta("runtimePropagation", before, after),
-        relation_complexity: signed_delta("relationComplexity", before, after),
-    }
-}
-
-fn metric_delta_status(
-    before: &SignatureSnapshot,
-    after: &SignatureSnapshot,
-) -> BTreeMap<String, MetricDeltaStatus> {
-    DATASET_DELTA_AXES
-        .iter()
-        .map(|axis| ((*axis).to_string(), delta_status(axis, before, after)))
-        .collect()
-}
-
-fn signed_delta(axis: &str, before: &SignatureSnapshot, after: &SignatureSnapshot) -> Option<i64> {
-    let before_measured = is_measured(axis, &before.metric_status);
-    let after_measured = is_measured(axis, &after.metric_status);
-    if !before_measured || !after_measured {
-        return None;
-    }
-
-    Some(signature_value(axis, &after.signature)? - signature_value(axis, &before.signature)?)
-}
-
-fn delta_status(
-    axis: &str,
-    before: &SignatureSnapshot,
-    after: &SignatureSnapshot,
-) -> MetricDeltaStatus {
-    let before_measured = is_measured(axis, &before.metric_status);
-    let after_measured = is_measured(axis, &after.metric_status);
-    let before_value = signature_value(axis, &before.signature);
-    let after_value = signature_value(axis, &after.signature);
-    let comparable =
-        before_measured && after_measured && before_value.is_some() && after_value.is_some();
-    let reason = if comparable {
-        "measured before and after".to_string()
-    } else {
-        delta_unavailable_reason(
-            axis,
-            before_measured,
-            after_measured,
-            before_value.is_some(),
-            after_value.is_some(),
-            &before.metric_status,
-            &after.metric_status,
-        )
-    };
-
-    MetricDeltaStatus {
-        comparable,
-        reason,
-        before_measured,
-        after_measured,
-    }
-}
-
-fn is_measured(axis: &str, metric_status: &BTreeMap<String, MetricStatus>) -> bool {
-    metric_status
-        .get(axis)
-        .is_some_and(|status| status.measured)
-}
-
-fn signature_value(axis: &str, signature: &ArchitectureSignatureV1DatasetShape) -> Option<i64> {
-    let value = match axis {
-        "hasCycle" => Some(signature.has_cycle),
-        "sccMaxSize" => Some(signature.scc_max_size),
-        "maxDepth" => Some(signature.max_depth),
-        "fanoutRisk" => Some(signature.fanout_risk),
-        "boundaryViolationCount" => Some(signature.boundary_violation_count),
-        "abstractionViolationCount" => Some(signature.abstraction_violation_count),
-        "sccExcessSize" => Some(signature.scc_excess_size),
-        "maxFanout" => Some(signature.max_fanout),
-        "reachableConeSize" => Some(signature.reachable_cone_size),
-        "weightedSccRisk" => signature.weighted_scc_risk,
-        "projectionSoundnessViolation" => signature.projection_soundness_violation,
-        "lspViolationCount" => signature.lsp_violation_count,
-        "nilpotencyIndex" => signature.nilpotency_index,
-        "runtimePropagation" => signature.runtime_propagation,
-        "relationComplexity" => signature.relation_complexity,
-        _ => None,
-    };
-    value.map(|value| value as i64)
-}
-
-fn delta_unavailable_reason(
-    axis: &str,
-    before_measured: bool,
-    after_measured: bool,
-    before_value_present: bool,
-    after_value_present: bool,
-    before_status: &BTreeMap<String, MetricStatus>,
-    after_status: &BTreeMap<String, MetricStatus>,
-) -> String {
-    let mut reasons = Vec::new();
-    if !before_measured {
-        reasons.push(format!(
-            "before: {}",
-            metric_status_reason(axis, before_status)
-        ));
-    }
-    if !after_measured {
-        reasons.push(format!(
-            "after: {}",
-            metric_status_reason(axis, after_status)
-        ));
-    }
-    if before_measured && !before_value_present {
-        reasons.push("before signature value is null".to_string());
-    }
-    if after_measured && !after_value_present {
-        reasons.push("after signature value is null".to_string());
-    }
-    reasons.join("; ")
-}
-
-fn metric_status_reason(axis: &str, metric_status: &BTreeMap<String, MetricStatus>) -> String {
-    metric_status
-        .get(axis)
-        .and_then(|status| status.reason.clone())
-        .unwrap_or_else(|| "metricStatus entry is missing".to_string())
-}
-
-fn snapshot_validation_summary(
-    report: Option<&ComponentUniverseValidationReport>,
-    report_path: Option<String>,
-) -> SnapshotValidationSummary {
-    match report {
-        Some(report) => SnapshotValidationSummary {
-            schema_version: Some(report.schema_version.clone()),
-            result: report.summary.result.clone(),
-            universe_mode: Some(report.universe_mode.clone()),
-            failed_check_count: Some(report.summary.failed_check_count),
-            warning_check_count: Some(report.summary.warning_check_count),
-            not_measured_check_count: Some(report.summary.not_measured_check_count),
-            report_path,
-        },
-        None => SnapshotValidationSummary {
-            schema_version: None,
-            result: "not_run".to_string(),
-            universe_mode: None,
-            failed_check_count: None,
-            warning_check_count: None,
-            not_measured_check_count: None,
-            report_path: None,
-        },
-    }
-}
-
-fn store_record_signature_snapshot(
-    record: &SignatureSnapshotStoreRecordV0,
-    role: &str,
-) -> SignatureSnapshot {
-    SignatureSnapshot {
-        commit: GitCommitRef {
-            sha: record.revision.sha.clone(),
-            role: role.to_string(),
-        },
-        extractor: ExtractorRef {
-            name: record.extractor.name.clone(),
-            version: record.extractor.version.clone(),
-            rule_set_version: record.extractor.rule_set_version.clone(),
-            policy_version: record.policy.policy_id.clone(),
-        },
-        signature: record.signature.clone(),
-        metric_status: record.metric_status.clone(),
-    }
-}
-
-fn snapshot_diff_endpoint(record: &SignatureSnapshotStoreRecordV0) -> SnapshotDiffEndpoint {
-    SnapshotDiffEndpoint {
-        repository: record.repository.clone(),
-        revision: record.revision.clone(),
-        validation_result: record.validation_summary.result.clone(),
-        extractor_version: record.extractor.version.clone(),
-        rule_set_version: record.extractor.rule_set_version.clone(),
-        policy_id: record.policy.policy_id.clone(),
-    }
-}
-
-fn snapshot_comparison_status(
-    before: &SignatureSnapshotStoreRecordV0,
-    after: &SignatureSnapshotStoreRecordV0,
-) -> SnapshotComparisonStatus {
-    let mut reasons = Vec::new();
-    if !matches!(before.validation_summary.result.as_str(), "pass" | "warn") {
-        reasons.push(format!(
-            "before validationSummary.result = {}",
-            before.validation_summary.result
-        ));
-    }
-    if !matches!(after.validation_summary.result.as_str(), "pass" | "warn") {
-        reasons.push(format!(
-            "after validationSummary.result = {}",
-            after.validation_summary.result
-        ));
-    }
-    if before.extractor.name != after.extractor.name {
-        reasons.push("extractor.name differs".to_string());
-    }
-    if before.extractor.version != after.extractor.version {
-        reasons.push("extractor.version differs".to_string());
-    }
-    if before.extractor.rule_set_version != after.extractor.rule_set_version {
-        reasons.push("extractor.ruleSetVersion differs".to_string());
-    }
-    if before.policy.policy_id != after.policy.policy_id {
-        reasons.push("policy.policyId differs".to_string());
-    }
-    if before.policy.version != after.policy.version {
-        reasons.push("policy.version differs".to_string());
-    }
-
-    SnapshotComparisonStatus {
-        primary_diff_eligible: reasons.is_empty(),
-        reasons,
-    }
-}
-
-fn classify_axis_deltas(
-    before: &SignatureSnapshotStoreRecordV0,
-    after: &SignatureSnapshotStoreRecordV0,
-    status: &BTreeMap<String, MetricDeltaStatus>,
-) -> (
-    Vec<SignatureAxisChange>,
-    Vec<SignatureAxisChange>,
-    Vec<SignatureAxisChange>,
-    Vec<UnmeasuredAxisDelta>,
-) {
-    let mut worsened = Vec::new();
-    let mut improved = Vec::new();
-    let mut unchanged = Vec::new();
-    let mut unmeasured = Vec::new();
-
-    for axis in DATASET_DELTA_AXES {
-        let axis_status = status
-            .get(axis)
-            .expect("metric delta status exists for every dataset delta axis");
-        let before_value = signature_value(axis, &before.signature);
-        let after_value = signature_value(axis, &after.signature);
-        if !axis_status.comparable {
-            unmeasured.push(UnmeasuredAxisDelta {
-                axis: axis.to_string(),
-                reason: axis_status.reason.clone(),
-                before_measured: axis_status.before_measured,
-                after_measured: axis_status.after_measured,
-            });
-            continue;
-        }
-
-        let before_value = before_value.expect("comparable before value exists");
-        let after_value = after_value.expect("comparable after value exists");
-        let change = SignatureAxisChange {
-            axis: axis.to_string(),
-            before: before_value,
-            after: after_value,
-            delta: after_value - before_value,
-        };
-        match change.delta.cmp(&0) {
-            std::cmp::Ordering::Greater => worsened.push(change),
-            std::cmp::Ordering::Less => improved.push(change),
-            std::cmp::Ordering::Equal => unchanged.push(change),
-        }
-    }
-
-    (worsened, improved, unchanged, unmeasured)
-}
-
-fn snapshot_evidence_diff(
-    before: Option<&Sig0Document>,
-    after: Option<&Sig0Document>,
-) -> SnapshotEvidenceDiff {
-    let (before, after) = match (before, after) {
-        (Some(before), Some(after)) => (before, after),
-        _ => {
-            return SnapshotEvidenceDiff {
-                available: false,
-                unavailable_reason: Some(
-                    "before and after ArchSig JSON are required for component / edge evidence diff"
-                        .to_string(),
-                ),
-                component_delta: None,
-                edge_delta: None,
-                policy_violation_delta: None,
-            };
-        }
-    };
-
-    let before_components = before
-        .components
-        .iter()
-        .map(|component| component.id.clone())
-        .collect::<BTreeSet<_>>();
-    let after_components = after
-        .components
-        .iter()
-        .map(|component| component.id.clone())
-        .collect::<BTreeSet<_>>();
-    let before_edges = before.edges.iter().cloned().collect::<BTreeSet<_>>();
-    let after_edges = after.edges.iter().cloned().collect::<BTreeSet<_>>();
-    let before_policy_violations = before
-        .policy_violations
-        .iter()
-        .cloned()
-        .collect::<BTreeSet<_>>();
-    let after_policy_violations = after
-        .policy_violations
-        .iter()
-        .cloned()
-        .collect::<BTreeSet<_>>();
-
-    SnapshotEvidenceDiff {
-        available: true,
-        unavailable_reason: None,
-        component_delta: Some(ComponentSetDelta {
-            before_count: before_components.len(),
-            after_count: after_components.len(),
-            delta: after_components.len() as i64 - before_components.len() as i64,
-            added: set_added(&before_components, &after_components),
-            removed: set_removed(&before_components, &after_components),
-        }),
-        edge_delta: Some(EdgeSetDelta {
-            before_count: before_edges.len(),
-            after_count: after_edges.len(),
-            delta: after_edges.len() as i64 - before_edges.len() as i64,
-            added: set_added(&before_edges, &after_edges),
-            removed: set_removed(&before_edges, &after_edges),
-        }),
-        policy_violation_delta: Some(PolicyViolationSetDelta {
-            before_count: before_policy_violations.len(),
-            after_count: after_policy_violations.len(),
-            delta: after_policy_violations.len() as i64 - before_policy_violations.len() as i64,
-            added: set_added(&before_policy_violations, &after_policy_violations),
-            removed: set_removed(&before_policy_violations, &after_policy_violations),
-        }),
-    }
-}
-
-fn set_added<T: Ord + Clone>(before: &BTreeSet<T>, after: &BTreeSet<T>) -> Vec<T> {
-    after.difference(before).cloned().collect()
-}
-
-fn set_removed<T: Ord + Clone>(before: &BTreeSet<T>, after: &BTreeSet<T>) -> Vec<T> {
-    before.difference(after).cloned().collect()
-}
-
-fn snapshot_diff_attribution(
-    after: &SignatureSnapshotStoreRecordV0,
-    evidence_diff: &SnapshotEvidenceDiff,
-    worsened_axes: &[SignatureAxisChange],
-    pr_metadata: &[EmpiricalDatasetInput],
-) -> SnapshotDiffAttribution {
-    let touched_components = touched_components_from_evidence_diff(evidence_diff);
-    let worsened_axis_names = worsened_axes
-        .iter()
-        .map(|axis| axis.axis.clone())
-        .collect::<Vec<_>>();
-    let mut candidates = pr_metadata
-        .iter()
-        .map(|metadata| {
-            attribution_candidate(
-                after,
-                evidence_diff,
-                &touched_components,
-                &worsened_axis_names,
-                metadata,
-            )
-        })
-        .collect::<Vec<_>>();
-    candidates.sort_by(|left, right| {
-        right
-            .confidence
-            .partial_cmp(&left.confidence)
-            .unwrap_or(std::cmp::Ordering::Equal)
-            .then(left.id.cmp(&right.id))
-    });
-
-    let method = if pr_metadata.is_empty() {
-        "none: PR metadata not provided".to_string()
-    } else if touched_components.is_empty() {
-        "pr-metadata: commit match and changed components; raw evidence diff unavailable or empty"
-            .to_string()
-    } else {
-        "pr-metadata: commit match and changed-component overlap with added/removed evidence"
-            .to_string()
-    };
-
-    let shared_worsened_axes = if candidates
-        .iter()
-        .filter(|candidate| candidate.confidence_level != "unknown")
-        .count()
-        > 1
-    {
-        worsened_axis_names
-    } else {
-        Vec::new()
-    };
-
-    SnapshotDiffAttribution {
-        method,
-        shared_worsened_axes,
-        candidates,
-    }
-}
-
-fn touched_components_from_evidence_diff(evidence_diff: &SnapshotEvidenceDiff) -> BTreeSet<String> {
-    let mut components = BTreeSet::new();
-    if let Some(component_delta) = &evidence_diff.component_delta {
-        components.extend(component_delta.added.iter().cloned());
-        components.extend(component_delta.removed.iter().cloned());
-    }
-    if let Some(edge_delta) = &evidence_diff.edge_delta {
-        for edge in edge_delta.added.iter().chain(edge_delta.removed.iter()) {
-            components.insert(edge.source.clone());
-            components.insert(edge.target.clone());
-        }
-    }
-    if let Some(policy_delta) = &evidence_diff.policy_violation_delta {
-        for violation in policy_delta.added.iter().chain(policy_delta.removed.iter()) {
-            components.insert(violation.source.clone());
-            components.insert(violation.target.clone());
-        }
-    }
-    components
-}
-
-fn attribution_candidate(
-    after: &SignatureSnapshotStoreRecordV0,
-    evidence_diff: &SnapshotEvidenceDiff,
-    touched_components: &BTreeSet<String>,
-    worsened_axes: &[String],
-    metadata: &EmpiricalDatasetInput,
-) -> AttributionCandidate {
-    let changed_components = metadata.pr_metrics.changed_components.clone();
-    let changed_set = changed_components.iter().cloned().collect::<BTreeSet<_>>();
-    let overlap = changed_set
-        .intersection(touched_components)
-        .cloned()
-        .collect::<Vec<_>>();
-    let matched_edges = matched_edges_from_evidence_diff(evidence_diff, &changed_set);
-    let matched_policy_violations =
-        matched_policy_violations_from_evidence_diff(evidence_diff, &changed_set);
-    let mut matched_components = overlap;
-    matched_components.sort();
-    matched_components.dedup();
-    let mut confidence = 0.0;
-    let mut reasons = Vec::new();
-
-    if after.revision.sha == metadata.pull_request.head_commit {
-        confidence += 0.35;
-        reasons.push("after revision matches pullRequest.headCommit".to_string());
-    }
-    if metadata
-        .pull_request
-        .merge_commit
-        .as_ref()
-        .is_some_and(|merge_commit| after.revision.sha == *merge_commit)
-    {
-        confidence += 0.35;
-        reasons.push("after revision matches pullRequest.mergeCommit".to_string());
-    }
-    if !matched_components.is_empty() {
-        let denominator = touched_components.len().max(1) as f64;
-        confidence += 0.35 * (matched_components.len() as f64 / denominator).min(1.0);
-        reasons.push(format!(
-            "changed components overlap with evidence diff: {}",
-            matched_components.join(", ")
-        ));
-    }
-    if !matched_edges.is_empty() {
-        confidence += 0.15;
-        reasons.push(format!(
-            "changed components touch added/removed edges: {}",
-            edge_labels(&matched_edges).join(", ")
-        ));
-    }
-    if !matched_policy_violations.is_empty() {
-        confidence += 0.15;
-        reasons.push(format!(
-            "changed components touch added/removed policy violations: {}",
-            policy_violation_labels(&matched_policy_violations).join(", ")
-        ));
-    }
-    if !worsened_axes.is_empty() && confidence > 0.0 {
-        confidence += 0.10;
-        reasons.push(format!(
-            "signature worsened on axes: {}",
-            worsened_axes.join(", ")
-        ));
-    }
-    if reasons.is_empty() {
-        reasons.push(
-            "weak candidate: PR metadata provided but no commit or evidence overlap".to_string(),
-        );
-    }
-    let confidence = round_confidence(confidence.min(0.95));
-
-    AttributionCandidate {
-        kind: "pullRequest".to_string(),
-        id: format!("#{}", metadata.pull_request.number),
-        confidence,
-        confidence_level: confidence_level(confidence).to_string(),
-        reasons,
-        changed_components,
-        matched_components,
-        matched_edges,
-        matched_policy_violations,
-        affected_axes: if confidence > 0.0 {
-            worsened_axes.to_vec()
-        } else {
-            Vec::new()
-        },
-    }
-}
-
-fn matched_edges_from_evidence_diff(
-    evidence_diff: &SnapshotEvidenceDiff,
-    changed_components: &BTreeSet<String>,
-) -> Vec<Edge> {
-    let Some(edge_delta) = &evidence_diff.edge_delta else {
-        return Vec::new();
-    };
-    edge_delta
-        .added
-        .iter()
-        .chain(edge_delta.removed.iter())
-        .filter(|edge| {
-            changed_components.contains(&edge.source) || changed_components.contains(&edge.target)
-        })
-        .cloned()
-        .collect()
-}
-
-fn matched_policy_violations_from_evidence_diff(
-    evidence_diff: &SnapshotEvidenceDiff,
-    changed_components: &BTreeSet<String>,
-) -> Vec<PolicyViolation> {
-    let Some(policy_delta) = &evidence_diff.policy_violation_delta else {
-        return Vec::new();
-    };
-    policy_delta
-        .added
-        .iter()
-        .chain(policy_delta.removed.iter())
-        .filter(|violation| {
-            changed_components.contains(&violation.source)
-                || changed_components.contains(&violation.target)
-        })
-        .cloned()
-        .collect()
-}
-
-fn edge_labels(edges: &[Edge]) -> Vec<String> {
-    edges
-        .iter()
-        .map(|edge| format!("{} -> {}", edge.source, edge.target))
-        .collect()
-}
-
-fn policy_violation_labels(violations: &[PolicyViolation]) -> Vec<String> {
-    violations
-        .iter()
-        .map(|violation| {
-            format!(
-                "{}:{} -> {}",
-                violation.axis, violation.source, violation.target
-            )
-        })
-        .collect()
-}
-
-fn confidence_level(confidence: f64) -> &'static str {
-    if confidence >= 0.80 {
-        "high"
-    } else if confidence >= 0.50 {
-        "medium"
-    } else if confidence > 0.0 {
-        "low"
-    } else {
-        "unknown"
-    }
-}
-
-fn round_confidence(value: f64) -> f64 {
-    (value * 100.0).round() / 100.0
-}
-
 fn count_checks(checks: &[ValidationCheck], result: &str) -> usize {
     checks.iter().filter(|check| check.result == result).count()
 }
@@ -4001,6 +2626,8 @@ fn unmeasured_status(reason: String) -> MetricStatus {
 
 #[cfg(test)]
 mod tests {
+    use std::path::Path;
+
     use super::*;
 
     #[test]
